Finalizing AOTF driver
Assuming the commands are correct, it should work to control the AOTF now. Still have to add a connection between the laser power sliders/laser objects and the AOTF object, as I have to control both lasers through the same AOTF object.
Double check also if \n or \r is the right termination of the commands (or try it out rather).