AbCorrCheck patterns added
Testing out to add the aberration correction patterns, used to figure out what aberrations I have in the system, in the master Tempesta, to try them out on the actual setup. Still have to add a way to rotate these, as this is the way I figure out what direction I have the aberrations in. Also read up on how I figure out how strong the aberrations are.