Add Imspector communication
This commit adds imspector communication to the tiling widget, so that it can now perform fully automatic tiling. It senses when the frames finish in imspector and then moves to the next tile with all that that entails, and works when scanning a Sync0-stack in Imspector for all the tiles!