Added two-color tiling
Tempesta now automatically checks if you are doing a one-color or two-color tiling stack in Imspector, and adapts to it. Also double checks that the tile size put in Tempesta agrees with that in Impspector.
In the future: maybe read also tile size just from Imspector, and that's it. Just decide the overlap in Tempesta basically.