Info saving for objectives
Added separate information saving for different objectives, so that the oil and glycerol objective both have their own information files. This makes it easier to go between them, as they might need slightly different parameters to be aligned. In this way you can load the latest parameters that was used for the specific objective you want to use.