Not working yet
Changed all "with bla bla as bla bla" in tempestaSLMKatana.py to just creating one of each, and that helped. Now it's complaining about certain methods in the SLM for example, but I am not sure this is due to the fact that I ruined something when I changed the first thing mentioned here, or if this is just an error from before due to not having the real SLM on.